RandomNumberGenerator fixes Float cast to Max rand
[git-p4: depot-paths = "//depot/C35/Students/cwsullivan/SD/": change = 424857]

+++ Engine/Code/Engine/Math/RandomNumberGenerator.cpp
@@ -13,7 +13,7 @@ int RandomNumberGenerator::RollRandomIntInRange(int minInclusive, int maxInclusi
 
 float RandomNumberGenerator::RollRandomFloatZeroToOne()
 {
-	return (float)rand() / RAND_MAX;
+	return (float)rand() / (float)RAND_MAX;
 }
 
 float RandomNumberGenerator::RollRandomFloatInRange(float minInclusive, float maxInclusive)
